+++ test/src/model/manager.spec.ts
@@ -1,6 +1,0 @@
-
-export function managerTests() {
-    describe('manager', () => {
-        //modelManagerTests();
-    });
-}